TEL-285: react compiler, server components, react 19, next 16
