+++ package.json
@@ -18,7 +18,7 @@
     "start": "next start"
   },
   "dependencies": {
-    "@clerk/nextjs": "^6.12.5",
+    "@clerk/nextjs": "6.37.1",
     "@prisma/client": "^6.4.1",
     "@supabase/supabase-js": "^2.45.4",
     "@t3-oss/env-nextjs": "^0.10.1",
@@ -33,10 +33,10 @@
     "classnames": "^2.5.1",
     "framer-motion": "^12.23.22",
     "geist": "^1.3.0",
-    "next": "15.5.9",
-    "npm": "^10.9.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "^16.1.6",
+    "npm": "10.9.0",
+    "react": "19.2.4",
+    "react-dom": "19.2.4",
     "react-hot-toast": "^2.5.2",
     "react-select": "^5.8.1",
     "react-spinners": "^0.15.0",
@@ -54,6 +54,7 @@
     "@types/react-dom": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^8.1.0",
     "@typescript-eslint/parser": "^8.1.0",
+    "babel-plugin-react-compiler": "^1.0.0",
     "eslint": "^8.57.0",
     "eslint-config-next": "15.5.9",
     "eslint-config-prettier": "^9.1.0",
