+++ src/app/_components/Buttons/BasicButton.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { type PropsWithChildren, memo, useState } from "react";
 
 import styles from "@/app/_components/Buttons/index.module.scss";
