+++ src/app/_components/Cars/CarScreenView.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import classNames from "classnames";
 import Image, { type StaticImageData } from "next/image";
 import { memo, useEffect, useState } from "react";
@@ -16,7 +18,6 @@ interface CarScreenViewProps {
   className: string | undefined;
   content: string;
   footerEnabled: boolean;
-  handleImageLoad: (() => void) | undefined;
   id: string;
   image: StaticImageData;
   navbarEnabled: boolean;
@@ -28,7 +29,6 @@ const CarScreenView = ({
   className,
   content,
   footerEnabled,
-  handleImageLoad,
   id,
   image,
   navbarEnabled,
@@ -88,7 +88,6 @@ const CarScreenView = ({
           className={styles.backSplashImage}
           fill
           loading="eager"
-          onLoad={handleImageLoad}
           placeholder="blur"
           priority
           src={image}
