+++ src/app/_components/OurWork/YearSection.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { AnimatePresence, LayoutGroup, motion } from "framer-motion";
 import Image from "next/image";
 import { memo, useCallback, useEffect, useRef, useState } from "react";
