+++ src/app/_components/Recruitment/index.module.scss
@@ -1,39 +1,35 @@
-.main {
-  min-height: 100dvh;
-  padding-bottom: 2rem;
-}
-
-.container {
+.application {
   display: flex;
   flex-direction: column;
-  flex-wrap: wrap;
-  justify-content: center;
   align-items: center;
-
-  padding: 0rem 12rem;
-
-  > img {
-    z-index: -1;
-  }
-
-  color: white;
-}
-
-.pageHeading {
-  width: 100%;
+  gap: 16px;
   text-align: center;
-  font-size: 5em;
-  color: white;
-  padding: 1em 0 0 0;
-  letter-spacing: 0.1em;
-  text-shadow: 0 0.5px 10px black;
-  font-weight: 450;
+
+  min-width: 20%;
+  max-width: 40%;
 }
 
 .applicationDescription {
   padding: 1em 0 3em 0;
   font-weight: 600;
   font-size: 1.2em;
+}
+
+.applicationContainer {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 48px;
+
+  .applicationTitle {
+    font-size: 1.8em;
+    font-weight: 600;
+  }
+
+  .applicationDescription {
+    font-size: 1.1em;
+    font-weight: 400;
+  }
 }
 
 @media (max-width: 1024px) {
@@ -43,6 +39,9 @@
   .pageHeading {
     letter-spacing: 0;
     padding-top: 0.8em;
+  }
+  .application {
+    min-width: 100%;
   }
 }
