+++ src/app/layout.tsx
@@ -21,7 +21,7 @@ export const metadata: Metadata = {
   title: "Calgary Solar Car",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: {
   children: React.ReactNode;
