--- src/app/sponsors/loading.tsx
@@ -1,9 +1,0 @@
-import { memo } from "react";
-
-import Loader from "@/app/_components/Loader";
-
-const Loading = () => {
-  return <Loader isLoading />;
-};
-
-export default memo(Loading);
