+++ src/app/sponsors/page.tsx
@@ -1,18 +1,14 @@
-"use client";
-
 import Image from "next/image";
 import Link from "next/link";
 import backsplash from "public/assets/sponsors/backsplash.jpeg";
-import { memo, useCallback, useState } from "react";
+import { memo } from "react";
 
 import Footer from "@/app/_components/Footer";
 import Navbar from "@/app/_components/Navbar";
 import styles from "@/app/sponsors/index.module.scss";
 import { type RouterOutputs } from "@/trpc/react";
-import { trpc } from "@/trpc/react";
+import { trpc } from "@/trpc/server";
 import { SponsorLevel } from "@prisma/client";
-
-import Loader from "../_components/Loader";
 
 type Sponsor = RouterOutputs["fe"]["getSponsors"][0];
 
@@ -45,18 +41,11 @@ const SponsorLevelImages = ({
   );
 };
 
-const Sponsors = () => {
-  const { data: sponsors, isFetching } = trpc.fe.getSponsors.useQuery();
-
-  const [isImageLoading, setIsImageLoading] = useState(true);
-
-  const handleImageLoad = useCallback(() => {
-    setIsImageLoading(false);
-  }, []);
+const Sponsors = async () => {
+  const sponsors = await trpc.fe.getSponsors();
 
   return (
     <>
-      {(isImageLoading || isFetching) && <Loader isLoading={isImageLoading} />}
       <main>
         <Navbar />
         <div className={styles.container}>
@@ -126,7 +115,6 @@ const Sponsors = () => {
             fill
             id="backsplashImage"
             loading="eager"
-            onLoad={handleImageLoad}
             placeholder="blur"
             priority
             src={backsplash}
@@ -138,5 +126,4 @@ const Sponsors = () => {
     </>
   );
 };
-
 export default memo(Sponsors);
