+++ src/app/support-us/page.tsx
@@ -1,25 +1,15 @@
-"use client";
-
 import Image from "next/image";
 import Link from "next/link";
 import backsplash from "public/assets/support-us/backsplash.jpeg";
-import { memo, useCallback, useState } from "react";
+import { memo } from "react";
 
 import Footer from "@/app/_components/Footer";
-import Loader from "@/app/_components/Loader";
 import Navbar from "@/app/_components/Navbar";
 import styles from "@/app/support-us/index.module.scss";
 
 const SupportUs = () => {
-  const [isImageLoading, setIsImageLoading] = useState(true);
-
-  const handleImageLoad = useCallback(() => {
-    setIsImageLoading(false);
-  }, []);
-
   return (
     <>
-      {isImageLoading && <Loader isLoading={isImageLoading} />}
       <main className={styles.main}>
         <Navbar />
         <div className={styles.container}>
@@ -42,7 +32,6 @@ const SupportUs = () => {
             fill
             id="backsplashImage"
             loading="eager"
-            onLoad={handleImageLoad}
             placeholder="blur"
             priority
             src={backsplash}
