+++ src/app/team/page.tsx
@@ -10,8 +10,6 @@ const Team = async () => {
   const teamHierarchy = await trpc.fe.getTeamMembers();
   const alumniTeam = await trpc.fe.getAlumni();
 
-  const tempFlag = true;
-
   const {
     accountingTeam,
     businessTeamManager,
@@ -42,152 +40,143 @@ const Team = async () => {
           working on our sixth car and have completed 7 races with previous
           vehicles.`}
           </div>
-          {tempFlag ? (
-            <div className={styles.teamMembersContainer}>
+          <div className={styles.teamMembersContainer}>
+            <div className={styles.teamRoleContainer}>
+              <div className={styles.title}>Co-Chairs</div>
+              <div className={styles.teamManagers}>
+                <TeamMember user={engineeringTeamManager} />
+                <TeamMember user={teamCaptain} />
+                <TeamMember user={businessTeamManager} />
+              </div>
+            </div>
+            {managerRoles && managerRoles.length > 0 && (
               <div className={styles.teamRoleContainer}>
-                <div className={styles.title}>Co-Chairs</div>
-                <div className={styles.teamManagers}>
-                  <TeamMember user={engineeringTeamManager} />
-                  <TeamMember user={teamCaptain} />
-                  <TeamMember user={businessTeamManager} />
+                <div className={styles.title}>Subteam Managers</div>
+                <div className={styles.teamMembers}>
+                  {managerRoles.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
                 </div>
               </div>
-              {managerRoles && managerRoles.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Subteam Managers</div>
-                  <div className={styles.teamMembers}>
-                    {managerRoles.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
+            )}
+            {leadRoles && leadRoles.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Team Leads</div>
+                <div className={styles.teamMembers}>
+                  {leadRoles.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
                 </div>
-              )}
-              {leadRoles && leadRoles.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Team Leads</div>
-                  <div className={styles.teamMembers}>
-                    {leadRoles.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {accountingTeam && accountingTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Accounting Team</div>
-                  <div className={styles.teamMembers}>
-                    {accountingTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {commmunicationsTeam && commmunicationsTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Communications Team</div>
-                  <div className={styles.teamMembers}>
-                    {commmunicationsTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {sponsorshipTeam && sponsorshipTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Sponsorship Team</div>
-                  <div className={styles.teamMembers}>
-                    {sponsorshipTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {softwareTeam && softwareTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Software Team</div>
-                  <div className={styles.teamMembers}>
-                    {softwareTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {electricalTeam && electricalTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Electrical Team</div>
-                  <div className={styles.teamMembers}>
-                    {electricalTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {mechanicalTeam && mechanicalTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Mechanical Team</div>
-                  <div className={styles.teamMembers}>
-                    {mechanicalTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {multiTeam && multiTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Multi Team</div>
-                  <div className={styles.teamMembers}>
-                    {multiTeam.map((teamMember) => (
-                      <TeamMember
-                        key={teamMember.clerkUserId}
-                        user={teamMember}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-              {alumniTeam && alumniTeam.length > 0 && (
-                <div className={styles.teamRoleContainer}>
-                  <div className={styles.title}>Alumni</div>
-                  <div className={styles.teamMembers}>
-                    {alumniTeam.map((teamMember) => (
-                      <TeamMember key={teamMember.id} user={teamMember} />
-                    ))}
-                  </div>
-                </div>
-              )}
-            </div>
-          ) : (
-            <div className={styles.teamMembersContainer}>
-              <div className={styles.description}>
-                {`This page is currently being built. Please check back later for
-                more information.`}
               </div>
-            </div>
-          )}
+            )}
+            {accountingTeam && accountingTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Accounting Team</div>
+                <div className={styles.teamMembers}>
+                  {accountingTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {commmunicationsTeam && commmunicationsTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Communications Team</div>
+                <div className={styles.teamMembers}>
+                  {commmunicationsTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {sponsorshipTeam && sponsorshipTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Sponsorship Team</div>
+                <div className={styles.teamMembers}>
+                  {sponsorshipTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {softwareTeam && softwareTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Software Team</div>
+                <div className={styles.teamMembers}>
+                  {softwareTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {electricalTeam && electricalTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Electrical Team</div>
+                <div className={styles.teamMembers}>
+                  {electricalTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {mechanicalTeam && mechanicalTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Mechanical Team</div>
+                <div className={styles.teamMembers}>
+                  {mechanicalTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {multiTeam && multiTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Multi Team</div>
+                <div className={styles.teamMembers}>
+                  {multiTeam.map((teamMember) => (
+                    <TeamMember
+                      key={teamMember.clerkUserId}
+                      user={teamMember}
+                    />
+                  ))}
+                </div>
+              </div>
+            )}
+            {alumniTeam && alumniTeam.length > 0 && (
+              <div className={styles.teamRoleContainer}>
+                <div className={styles.title}>Alumni</div>
+                <div className={styles.teamMembers}>
+                  {alumniTeam.map((teamMember) => (
+                    <TeamMember key={teamMember.id} user={teamMember} />
+                  ))}
+                </div>
+              </div>
+            )}
+          </div>
         </div>
       </main>
       <Footer />
